uboot-sign: Fix to install nonexistent dtb file
Add to check dtb file exist, then install it.

+++ meta/classes-recipe/uboot-sign.bbclass
@@ -318,7 +318,9 @@ EOF
 			${SPL_MKIMAGE_SIGN_ARGS}
 	fi
 
-	cp ${SPL_DIR}/${SPL_DTB_BINARY} ${SPL_DIR}/${SPL_DTB_SIGNED}
+	if [ -e "${SPL_DIR}/${SPL_DTB_BINARY}" ]; then
+		cp ${SPL_DIR}/${SPL_DTB_BINARY} ${SPL_DIR}/${SPL_DTB_SIGNED}
+	fi
 }
 
 uboot_assemble_fitimage_helper() {
